+++ include/invio/KLTTracker.h
@@ -22,9 +22,9 @@ public:
 	KLTTracker();
 	virtual ~KLTTracker();
 
-	void findNewFeaturePositions(const Frame& lf, const Frame& cf, const std::vector<Eigen::Vector2d>& previous_feature_positions,
-			const std::list<Feature>& estimated_new_feature_positions, std::vector<Eigen::Vector2d>& measured_positions,
-			std::vector<Eigen::Matrix2d>& estimated_uncertainty, std::vector<bool>& passed);
+	void findNewFeaturePositions(const Frame& lf, const Frame& cf, const std::vector<Eigen::Vector2f>& previous_feature_positions,
+			const std::list<Feature>& estimated_new_feature_positions, std::vector<Eigen::Vector2f>& measured_positions,
+			std::vector<Eigen::Matrix2f>& estimated_uncertainty, std::vector<bool>& passed);
 };
 
 #endif /* INVIO_INCLUDE_INVIO_KLTTRACKER_H_ */
